api/urls.py
from django.urls import path
from .views import ProcessView, StatusView
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
)
urlpatterns = [
    path('process/', ProcessView.as_view()),
    path('status/<str:task_id>/', StatusView.as_view()),
    path('token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
]